requirements.txt
faiss-cpu
sentence-transformers
transformers
accelerate
EOF
